Moved RecoverableConnectorException back to former namespace.
