+++ src/Connector/RecoverableConnectorException.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace ScriptFUSION\Porter\Connector\Recoverable;
+namespace ScriptFUSION\Porter\Connector;
 
 /**
  * The exception that is thrown when a recoverable (non-fatal) error occurs during Connector::fetch.
